task 7

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -3,7 +3,7 @@
 DELIMITER // 
 CREATE PROCEDURE ComputeAverageScoreForUser(user_id INT)
 BEGIN
-  DECLARE new_average INT;
+  DECLARE new_average FLOAT;
   SELECT AVG(score) FROM corrections WHERE user_id = user_id INTO new_average;
   
   UPDATE  users SET average_score = new_average WHERE id = user_id;
